simplifications
- wip for slimmer save/load
- WeaponType struct added
- Complexity reduction

+++ ManagedDoom/src/Doom/Doom.cs
@@ -402,7 +402,7 @@ namespace ManagedDoom.Doom
                         if (sendPause)
                         {
                             sendPause = false;
-                            cmds[Options.ConsolePlayer].Buttons |= (byte)(TicCmdButtons.Special | TicCmdButtons.Pause);
+                            cmds[Options.ConsolePlayer].Buttons |= TicCmdButtons.Special | TicCmdButtons.Pause;
                         }
                         if (Game.Update(cmds) == UpdateResult.NeedWipe)
                             StartWipe();
